# openapi-comparison/API Blueprint/generated-spring/src/main/java/org/openapitools/api/BooksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import java.math.BigDecimal;
import org.openapitools.model.CreateANewBookRequest;
import org.openapitools.model.GetAllBooks200ResponseInner;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-04-03T08:59:41.342205400+07:00[Asia/Saigon]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "Books", description = "the Books API")
public interface BooksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_A_NEW_BOOK = "/books";
    /**
     * POST /books : Create a new book
     * 
     *
     * @param createANewBookRequest  (optional)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "createANewBook",
        summary = "Create a new book",
        description = "",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllBooks200ResponseInner.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BooksApi.PATH_CREATE_A_NEW_BOOK,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<GetAllBooks200ResponseInner> createANewBook(
        @Parameter(name = "CreateANewBookRequest", description = "") @Valid @RequestBody(required = false) @Nullable CreateANewBookRequest createANewBookRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"published_year\" : 6.027456183070403, \"author\" : \"author\", \"id\" : 0.8008281904610115, \"title\" : \"title\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DELETE_A_BOOK = "/books/{bookId}";
    /**
     * DELETE /books/{bookId} : Delete a book
     * 
     *
     * @param bookId  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "deleteABook",
        summary = "Delete a book",
        description = "",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK"),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = BooksApi.PATH_DELETE_A_BOOK
    )
    default ResponseEntity<Void> deleteABook(
        @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") BigDecimal bookId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_A_BOOK_BY_ID = "/books/{bookId}";
    /**
     * GET /books/{bookId} : Get a book by ID
     * 
     *
     * @param bookId  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "getABookByID",
        summary = "Get a book by ID",
        description = "",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllBooks200ResponseInner.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_GET_A_BOOK_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<GetAllBooks200ResponseInner> getABookByID(
        @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") BigDecimal bookId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"published_year\" : 6.027456183070403, \"author\" : \"author\", \"id\" : 0.8008281904610115, \"title\" : \"title\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_ALL_BOOKS = "/books";
    /**
     * GET /books : Get all books
     * 
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getAllBooks",
        summary = "Get all books",
        description = "",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = GetAllBooks200ResponseInner.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_GET_ALL_BOOKS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<GetAllBooks200ResponseInner>> getAllBooks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"published_year\" : 6.027456183070403, \"author\" : \"author\", \"id\" : 0.8008281904610115, \"title\" : \"title\" }, { \"published_year\" : 6.027456183070403, \"author\" : \"author\", \"id\" : 0.8008281904610115, \"title\" : \"title\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_A_BOOK = "/books/{bookId}";
    /**
     * PUT /books/{bookId} : Update a book
     * 
     *
     * @param bookId  (required)
     * @param createANewBookRequest  (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateABook",
        summary = "Update a book",
        description = "",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetAllBooks200ResponseInner.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = BooksApi.PATH_UPDATE_A_BOOK,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<GetAllBooks200ResponseInner> updateABook(
        @Parameter(name = "bookId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("bookId") BigDecimal bookId,
        @Parameter(name = "CreateANewBookRequest", description = "") @Valid @RequestBody(required = false) @Nullable CreateANewBookRequest createANewBookRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"published_year\" : 6.027456183070403, \"author\" : \"author\", \"id\" : 0.8008281904610115, \"title\" : \"title\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}